+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 Unreleased
 ----------
 
+* Added the ``NoopController`` controller, that simply lets requests through.
+
 * Improved the performance of the ``Scheduler`` by better
   detecting availability conditions for queued requests,
   thus no longer requiring an internal lock.